API: Renamed module ai to ai_integration; text_generation separate out as a feature like embeddings
