+++ src-tauri/src/ai_integration/commands.rs
