+++ src-tauri/src/ai_integration/configuration.rs
