+++ src-tauri/src/ai_integration/crud.rs
