+++ src-tauri/src/ai_integration/models.rs
