+++ src-tauri/src/ai_integration/providers/mod.rs
