+++ src-tauri/src/ai_integration/providers/openai.rs
