+++ src-tauri/src/ai_integration/tool_use.rs
